Update licensing information and enhance Pro features integration
- Updated LICENSE.md to clarify the MIT License applicability and introduce the React on Rails Pro License for specific directories.
- Refactored helper methods in `lib/react_on_rails/helper.rb` to utilize new Pro features, including immediate hydration support.
- Introduced `lib/react_on_rails/pro_features/helper.rb` to encapsulate Pro feature logic, including immediate hydration and warning badge generation.
- Adjusted imports in `node_package/src/ReactOnRails.node.ts` to reflect the new Pro directory structure.
- Added new files for React Server Component handling and RSC payload management in the Pro directory, enhancing server-side rendering capabilities.
- Improved error handling and stream management for RSC payloads to optimize performance and debugging.
