+++ LICENSE.md
@@ -5,7 +5,12 @@
 
 ---
 
-# MIT License
+## MIT License for Core React on Rails
+
+This license applies to all files within this repository, with the exception of the code located in the following directories, which are licensed separately under the React on Rails Pro License:
+
+- `lib/react_on_rails/pro_features/`
+- `node_package/src/pro/`
 
 Copyright (c) 2017, 2018 Justin Gordon and ShakaCode  
 Copyright (c) 2015–2025 ShakaCode, LLC
@@ -31,3 +36,12 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
+
+---
+
+## React on Rails Pro License
+
+The code in the directories listed above is part of the React on Rails Pro framework and is licensed under the React on Rails Pro License.
+
+You can find the full text of the license agreement here:
+[REACT-ON-RAILS-PRO-LICENSE.md](./REACT-ON-RAILS-PRO-LICENSE.md)
