+++ node_package/src/pro/RSCProvider.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import type { ClientGetReactServerComponentProps } from './getReactServerComponent.client.ts';
-import { createRSCPayloadKey } from './utils.ts';
+import { createRSCPayloadKey } from '../utils.ts';
 
 type RSCContextType = {
   getComponent: (componentName: string, componentProps: unknown) => Promise<React.ReactNode>;
