+++ node_package/src/pro/getReactServerComponent.client.ts
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { createFromReadableStream } from 'react-on-rails-rsc/client.browser';
-import { createRSCPayloadKey, fetch, wrapInNewPromise, extractErrorMessage } from './utils.ts';
+import { createRSCPayloadKey, fetch, wrapInNewPromise, extractErrorMessage } from '../utils.ts';
 import transformRSCStreamAndReplayConsoleLogs from './transformRSCStreamAndReplayConsoleLogs.ts';
-import { RailsContext } from './types/index.ts';
+import { RailsContext } from '../types/index.ts';
 
 declare global {
   interface Window {
