+++ node_package/src/pro/getReactServerComponent.server.ts
@@ -1,8 +1,8 @@
 import { BundleManifest } from 'react-on-rails-rsc';
 import { buildClientRenderer } from 'react-on-rails-rsc/client.node';
 import transformRSCStream from './transformRSCNodeStream.ts';
-import loadJsonFile from './loadJsonFile.ts';
-import type { RailsContextWithServerStreamingCapabilities } from './types/index.ts';
+import loadJsonFile from '../loadJsonFile.ts';
+import type { RailsContextWithServerStreamingCapabilities } from '../types/index.ts';
 
 type GetReactServerComponentOnServerProps = {
   componentName: string;
