+++ node_package/src/pro/injectRSCPayload.ts
@@ -1,7 +1,7 @@
 import { PassThrough } from 'stream';
 import { finished } from 'stream/promises';
-import { createRSCPayloadKey } from './utils.ts';
-import { PipeableOrReadableStream } from './types/index.ts';
+import { createRSCPayloadKey } from '../utils.ts';
+import { PipeableOrReadableStream } from '../types/index.ts';
 import RSCRequestTracker from './RSCRequestTracker.ts';
 
 // In JavaScript, when an escape sequence with a backslash (\) is followed by a character
