+++ node_package/src/pro/registerServerComponent/server.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import ReactOnRails from '../ReactOnRails.client.ts';
+import ReactOnRails from '../../ReactOnRails.client.ts';
 import RSCRoute from '../RSCRoute.tsx';
-import { ReactComponent, RenderFunction } from '../types/index.ts';
+import { ReactComponent, RenderFunction } from '../../types/index.ts';
 import wrapServerComponentRenderer from '../wrapServerComponentRenderer/server.tsx';
 
 /**
