+++ node_package/src/pro/streamServerRenderedReactComponent.ts
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import { PassThrough, Readable } from 'stream';
 
-import * as ComponentRegistry from './pro/ComponentRegistry.ts';
-import createReactOutput from './createReactOutput.ts';
-import { isPromise, isServerRenderHash } from './isServerRenderResult.ts';
-import buildConsoleReplay from './buildConsoleReplay.ts';
-import handleError from './handleError.ts';
-import { renderToPipeableStream } from './ReactDOMServer.cts';
-import { createResultObject, convertToError, validateComponent } from './serverRenderUtils.ts';
+import * as ComponentRegistry from './ComponentRegistry.ts';
+import createReactOutput from '../createReactOutput.ts';
+import { isPromise, isServerRenderHash } from '../isServerRenderResult.ts';
+import buildConsoleReplay from '../buildConsoleReplay.ts';
+import handleError from '../handleError.ts';
+import { renderToPipeableStream } from '../ReactDOMServer.cjs';
+import { createResultObject, convertToError, validateComponent } from '../serverRenderUtils.ts';
 import {
   assertRailsContextWithServerStreamingCapabilities,
   RenderParams,
@@ -16,7 +16,7 @@ import {
   PipeableOrReadableStream,
   RailsContextWithServerStreamingCapabilities,
   assertRailsContextWithServerComponentMetadata,
-} from './types/index.ts';
+} from '../types/index.ts';
 import injectRSCPayload from './injectRSCPayload.ts';
 import PostSSRHookTracker from './PostSSRHookTracker.ts';
 import RSCRequestTracker from './RSCRequestTracker.ts';
