+++ node_package/src/pro/wrapServerComponentRenderer/client.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import * as ReactDOMClient from 'react-dom/client';
-import { ReactComponentOrRenderFunction, RenderFunction } from '../types/index.ts';
-import isRenderFunction from '../isRenderFunction.ts';
-import { ensureReactUseAvailable } from '../reactApis.cts';
+import { ReactComponentOrRenderFunction, RenderFunction } from '../../types/index.ts';
+import isRenderFunction from '../../isRenderFunction.ts';
+import { ensureReactUseAvailable } from '../../reactApis.cjs';
 import { createRSCProvider } from '../RSCProvider.tsx';
 import getReactServerComponent from '../getReactServerComponent.client.ts';
