+++ node_package/src/pro/wrapServerComponentRenderer/server.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-import type { RenderFunction, ReactComponentOrRenderFunction } from '../types/index.ts';
+import type { RenderFunction, ReactComponentOrRenderFunction } from '../../types/index.ts';
 import getReactServerComponent from '../getReactServerComponent.server.ts';
 import { createRSCProvider } from '../RSCProvider.tsx';
-import isRenderFunction from '../isRenderFunction.ts';
-import { assertRailsContextWithServerStreamingCapabilities } from '../types/index.ts';
+import isRenderFunction from '../../isRenderFunction.ts';
+import { assertRailsContextWithServerStreamingCapabilities } from '../../types/index.ts';
 
 /**
  * Wraps a client component with the necessary RSC context and handling for server-side operations.
